+++ src/entity/Stray.php
