+++ src/entity/animals/Land/Fox.php
