+++ src/entity/animals/Land/Mooshroom.php
