+++ src/entity/animals/Land/Mule.php
