+++ src/entity/animals/Land/Panda.php
