+++ src/entity/animals/Land/Rabbit.php
