+++ src/entity/animals/Water/Pufferfish.php
@@ -1,0 +1,54 @@
+<?php
+
+/*
+ *
+ *  ____            _        _   __  __ _                  __  __ ____
+ * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \
+ * | |_) / _ \ / __| |/ / _ \ __| |\/| | | '_ \ / _ \_____| |\/| | |_) |
+ * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/
+ * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_|
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * @author PocketMine Team
+ * @link http://www.pocketmine.net/
+ *
+ *
+ */
+
+declare(strict_types=1);
+
+namespace pocketmine\entity\animals;
+
+use pocketmine\event\entity\EntityDamageByEntityEvent;
+use pocketmine\event\entity\EntityDamageEvent;
+use pocketmine\entity\EntitySizeInfo;
+use pocketmine\item\VanillaItems;
+use pocketmine\math\Vector3;
+use pocketmine\nbt\tag\CompoundTag;
+use pocketmine\network\mcpe\protocol\types\entity\EntityIds;
+use function atan2;
+use function mt_rand;
+use function sqrt;
+use const M_PI;
+
+class Pufferfish extends WaterAnimal{
+
+	public static function getNetworkTypeId() : string{ return EntityIds::PUFFERFISH; }
+
+	protected function getInitialSizeInfo() : EntitySizeInfo{ return new EntitySizeInfo(0.95, 0.95); }
+
+	public function initEntity(CompoundTag $nbt) : void{
+		$this->setMaxHealth(10);
+		parent::initEntity($nbt);
+	}
+
+	public function getName() : string{
+		return "Pufferfish";
+	}
+
+	public function getDrops() : array{}
+}
