+++ src/entity/animals/Water/Tadpole.php
@@ -21,31 +21,34 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\animals;
 
+use pocketmine\event\entity\EntityDamageByEntityEvent;
 use pocketmine\event\entity\EntityDamageEvent;
-use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataCollection;
-use pocketmine\network\mcpe\protocol\types\entity\EntityMetadataFlags;
+use pocketmine\entity\EntitySizeInfo;
+use pocketmine\item\VanillaItems;
+use pocketmine\math\Vector3;
+use pocketmine\nbt\tag\CompoundTag;
+use pocketmine\network\mcpe\protocol\types\entity\EntityIds;
+use function atan2;
+use function mt_rand;
+use function sqrt;
+use const M_PI;
 
-abstract class WaterAnimal extends Living implements Ageable{
-	/** @var bool */
-	protected $baby = false;
+class Tadpole extends WaterAnimal{
 
-	public function isBaby() : bool{
-		return $this->baby;
+	public static function getNetworkTypeId() : string{ return EntityIds::TADPOLE; }
+
+	protected function getInitialSizeInfo() : EntitySizeInfo{ return new EntitySizeInfo(0.95, 0.95); }
+
+	public function initEntity(CompoundTag $nbt) : void{
+		$this->setMaxHealth(10);
+		parent::initEntity($nbt);
 	}
 
-	public function canBreathe() : bool{
-		return $this->isUnderwater();
+	public function getName() : string{
+		return "Tadpole";
 	}
 
-	public function onAirExpired() : void{
-		$ev = new EntityDamageEvent($this, EntityDamageEvent::CAUSE_SUFFOCATION, 2);
-		$this->attack($ev);
-	}
-
-	protected function syncNetworkData(EntityMetadataCollection $properties) : void{
-		parent::syncNetworkData($properties);
-		$properties->setGenericFlag(EntityMetadataFlags::BABY, $this->baby);
-	}
+	public function getDrops() : array{}
 }
